# src/api.yaml
openapi: 3.0.0
info:
  title: Employee management system API
  description: Api Testing.
  version: 1.0 
servers:
  - url: http://localhost:8000/
    description:  Local server 
  # - url: https://prod.com/
  #   description:  Pre Production server
  # - url: https://test.com/
  #   description:  Production server
components:
  securitySchemes:
    token:        # arbitrary name for the security scheme
          
      type: http
      scheme: bearer
    
    # ApiKey:        # arbitrary name for the security scheme
    #   type: apiKey
    #   in: header       # can be "header", "query" or "cookie"
    #   name: apikey

# ----------------------------------- Authentication swagger api --------------------------- #

#login api
paths:
...
  /api/user:
    post:
      tags:
        - user
      security:
       - token: []
      summary: Returns a success message.        
      post:
      requestBody:
        required: true
        content:
          # multipart/form-data:
           application/json:
            schema:
              properties: 
                employee_id:         
                  type: String
                  example : "D9-1"
                first_name:       
                  type: string
                  example : "abc"
                last_name:       
                  type: string
                  example : "xyz"
                email:       
                  type: string
                  example : "xyz@gmail.com"
                password:       
                  type: string
                  example : ""
                phone:
                  type: integer
                  example : "1234567890"
                report_by:       
                  type: string
                  example : ""
                role_id:       
                  type: string
                  example : ""
                department_id:       
                  type: string
                  example : ""
                designation_id:   
                  type: string
                  example : ""
                joining_date:       
                    type: string
                    format: date
                    example : "2023-09-01"
                status:
                    type: "string"
                    enum: ['Active', 'Inactive']
              required:
                - employee_id
                - first_name
                - last_name
                - email
                - phone
                - report_by
                - joining_date
                - status
                - password
                - role_id
                - department_id
                - designation_id
         
      responses:
        '201':
          description: Success message. 
        '400':
          description: Bad Request.
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description: Server error.

# all user get
...
  /api/user/:
    get:
      tags: ["user"]
      security:
       - token: []
      summary: Returns a user details.
      responses: 
        '200':
          description: Success message
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description:  Server error.

# get single user
...
  /api/user/{id}:
    get:
      tags: ["user"]
      security:
       - token: []
      summary: Returns a user details by ID.
      parameters:
        - name: id
          in: path
          required: true
          description: Parameter description in CommonMark or HTML.
      responses: 
        '200':
          description: Success message
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description:  Server error.
 #  update user
    put:
      tags: ["user"]
      security:
       - token: []
      summary: user detail update.
      parameters:
        - name: id
          in: path
          required: true
          description: Parameter description in CommonMark or HTML.
      put:
      requestBody:
        required: true
        content:
          # multipart/form-data:
           application/json:
            schema:
              properties: 
                address:         
                  type: String
                  example : ""

      responses: 
        '200':
          description: Success message
        '400':
          description: Bad Request
        '404':
          description: Not found
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description:  Server error.

 #  delete user
    delete:
      tags: ["user"]
      security:
       - token: []
      summary: user detail delete.
      parameters:
        - name: id
          in: path
          required: true
          description: Parameter description in CommonMark or HTML.
  
      responses: 
        '200':
          description: Success message
        '400':
          description: Bad Request
        '404':
          description: Not found
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description:  Server error.
 #  update status user
    patch:
      tags: ["user"]
      security:
       - token: []
      summary: user status update.
      parameters:
        - name: id
          in: path
          required: true
          description: Parameter description in CommonMark or HTML.
  
      responses: 
        '200':
          description: Success message
        '404':
          description: Not found
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description:  Server error.

#  check email
...
  /api/user/email:
    post:
      tags:
        - user
      security:
       - token: []
      summary: check email.     
      post:
      requestBody:
        required: true
        content:
          # multipart/form-data:
           application/json:
            schema:
              properties: 
                email:       
                  type: string
                  example : "xyz@gmail.com"
              required:
                - email
         
      responses:
        '200':
          description: Success message. 
        '400':
          description: Bad Request.
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description: Server error.

#  check employee id
...
  /api/user/employeeId:
    post:
      tags:
        - user
      security:
       - token: []
      summary: check employeeId.     
      post:
      requestBody:
        required: true
        content:
          # multipart/form-data:
           application/json:
            schema:
              properties: 
                employee_id:       
                  type: string
                  example : "D9-1"
              required:
                - email
         
      responses:
        '200':
          description: Success message. 
        '400':
          description: Bad Request.
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description: Server error.

#  get login info.
...
  /api/user/loginInfo:
    post:
      tags:
        - user
      security:
       - token: []
      summary:  return login info by id.     
      post:
      requestBody:
        required: true
        content:
          # multipart/form-data:
           application/json:
            schema:
              properties: 
                id:       
                  type: string
                  example : "6503e8ac937cf7d6982ba97f"
                startDate:       
                  type: string
                  example : "2023-09-19T00:00:00.999Z"
                endDate:       
                  type: string
                  example : "2023-09-20T18:29:59.999Z"
              required:
                - id
                - startDate
                - endDate
         
      responses:
        '200':
          description: Success message. 
        '400':
          description: Bad Request.
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description: Server error.

#  get USERNAME.
...
  /api/user/username:
    post:
      tags:
        - user
      security:
       - token: []
      summary:  return username     
         
      responses:
        '200':
          description: Success message. 
        '400':
          description: Bad Request.
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description: Server error.


##--------------------------------------------  user profile api -----------------------------------------##
# change image
...
  /api/user/image:
    post:
      tags:
        - Profile
      security:
       - token: []
      summary:  change image.     
      post:
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                profile_image:
                  type: string
                  format: binary
         
      responses:
        '200':
          description: Success message. 
        '400':
          description: Bad Request.
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description: Server error.

#  change password
...
  /api/user/password:
    post:
      tags:
        - Profile
      security:
       - token: []
      summary: Change password.        
      post:
      requestBody:
        required: true
        content:
          # multipart/form-data:
           application/json:
            schema:
              properties: 
                current_password:         
                  type: String
                  example : ""
                new_password:       
                  type: string
                  example : ""
                confirm_password:       
                  type: string
                  example : ""
              required:
                - current_password
                - new_password
                - confirm_password
         
      responses:
        '200':
          description: Success message. 
        '400':
          description: Bad Request.
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description: Server error.

##--------------------------------------------  account api -----------------------------------------##
...
  /api/user_document:
    post:
      tags:
        - User document
      security:
       - token: []
      summary:  upload document for user.     
      post:
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                resume:
                  type: string
                  format: binary
                offer_letter:
                  type: string
                  format: binary
                joining_letter:
                  type: string
                  format: binary
                other:
                  type: string
                  format: binary
                user_id:
                  type: string
         
      responses:
        '200':
          description: Success message. 
        '201':
          description: created successfully. 
        '400':
          description: Bad Request.
        '404':
          description: Not Found.
        '401':
          description:  Unauthorized.
        '403':
          description: Access Denied.
        '500':
          description: Server error.

##--------------------------------------------  emergency contact api -----------------------------------------##
...